+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2019-09-08
+
 ### Added
 
 - **[New Refactoring]** Bubble up If Statement
@@ -170,7 +172,8 @@ function doSomething(someData) {
 - **[New refactoring]** Extract Variable
 - **[New refactoring]** Rename Symbol
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/0.5.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/nicoespeon/abracadabra/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/nicoespeon/abracadabra/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/nicoespeon/abracadabra/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/nicoespeon/abracadabra/compare/0.2.0...0.3.0